requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
atari-py=0.2.6=pypi_0
blas=1.0=mkl
box2d-kengz=2.3.3=pypi_0
box2d-py=2.3.8=pypi_0
ca-certificates=2020.10.14=0
certifi=2020.11.8=py36h06a4308_0
cffi=1.14.4=pypi_0
chardet=3.0.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cython=3.0a6=pypi_0
dataclasses=0.7=py36_0
fasteners=0.15=pypi_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
glfw=2.0.0=pypi_0
gym=0.9.6=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
jsonpatch=1.27=pypi_0
jsonpointer=2.0=pypi_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.2=heb0550a_3
mkl=2020.2=256
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
monotonic=1.5=pypi_0
mujoco-py=2.0.2.13=pypi_0
ncurses=6.2=he6710b0_1
ninja=1.10.1=py36hfd86e86_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openssl=1.1.1h=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=8.0.1=py36he98fc37_0
pip=20.2.4=py36h06a4308_0
psutil=5.7.3=pypi_0
pycparser=2.20=pypi_0
pyglet=1.3.2=pypi_0
pyopengl=3.1.5=pypi_0
python=3.6.12=hcff3b4d_2
pytorch=1.7.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pyzmq=20.0.0=pypi_0
readline=8.0=h7b6447c_0
requests=2.25.0=pypi_0
scipy=1.5.4=pypi_0
setuptools=50.3.1=py36h06a4308_1
six=1.15.0=py36h06a4308_0
sqlite=3.33.0=h62c20be_0
swig=3.0.12=h38cdd7d_3
tk=8.6.10=hbc83047_0
torchaudio=0.7.0=py36
torchfile=0.1.0=pypi_0
torchvision=0.8.1=py36_cu102
tornado=6.1=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pypi_0
visdom=0.1.8.9=pypi_0
websocket-client=0.57.0=pypi_0
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
